Refactor template function related code into new tmplfunc directory
Earlier commit created a tmplfunc directory for new template function. Move
other template function related code into new directory

Changes:
 - Move template_function.go (+ test file) into new directory. Rename to
 tmpl_func.go to match directory naming
 - Split out existing `hclServiceFunc` template function from tmpl_func.go into
 its own file hcl_service_func.go
 - Move healthService related code (used by hclServiceFunc) into
 hcl_service_func.go
 - Rename HCLTmplFuncMap => HCLMap since it is now in the tmplfunc directory
 (less redundant wording)
