+++ templates/tftmpl/tmplfunc/hcl_service_func_test.go
@@ -1,4 +1,4 @@
-package tftmpl
+package tmplfunc
 
 import (
 	"testing"
@@ -6,39 +6,6 @@ import (
 	"github.com/hashicorp/hcat/dep"
 	"github.com/stretchr/testify/assert"
 )
-
-func TestJoinStringsFunc(t *testing.T) {
-	testCases := []struct {
-		name     string
-		content  []string
-		expected string
-	}{
-		{
-			"empty",
-			[]string{},
-			"",
-		}, {
-			"string",
-			[]string{"foobar"},
-			"foobar",
-		}, {
-			"multiple strings",
-			[]string{"foo", "bar", "baz"},
-			"foo.bar.baz",
-		}, {
-			"empty string ignored",
-			[]string{"foo", "", "baz"},
-			"foo.baz",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			actual := joinStringsFunc(".", tc.content...)
-			assert.Equal(t, tc.expected, actual)
-		})
-	}
-}
 
 func TestHCLServiceFunc(t *testing.T) {
 	testCases := []struct {
