+++ templates/tftmpl/tmplfunc/tmpl_func.go
@@ -1,4 +1,4 @@
-package tftmpl
+package tmplfunc
 
 import (
 	"fmt"
@@ -9,13 +9,11 @@ import (
 	"github.com/hashicorp/hcat"
 	"github.com/hashicorp/hcat/dep"
 	"github.com/hashicorp/hcat/tfunc"
-	"github.com/hashicorp/hcl/v2/gohcl"
-	"github.com/hashicorp/hcl/v2/hclwrite"
 )
 
-// HCLTmplFuncMap is the map of template functions for rendering HCL
+// HCLMap is the map of template functions for rendering HCL
 // to their respective implementations
-func HCLTmplFuncMap(meta map[string]map[string]string) template.FuncMap {
+func HCLMap(meta map[string]map[string]string) template.FuncMap {
 	tmplFuncs := hcat.FuncMapConsulV1()
 	tmplFuncs["indent"] = tfunc.Helpers()["indent"]
 	tmplFuncs["subtract"] = tfunc.Math()["subtract"]
@@ -58,30 +56,5 @@ func hclServiceTagsFunc() func(tags *dep.ServiceTags) string {
 		sort.Strings(t)
 
 		return fmt.Sprintf("[%s]", strings.Join(t, ", "))
-	}
-}
-
-// hclServiceFunc is a wrapper of the template function to marshal Consul
-// service information into HCL. The function accepts a map representing
-// metadata for services in scope of a task.
-func hclServiceFunc(meta map[string]map[string]string) func(sDep *dep.HealthService) string {
-	return func(sDep *dep.HealthService) string {
-		if sDep == nil {
-			return ""
-		}
-
-		// Find metdata based Consul service name scoped to a task to append to
-		// that service within var.services
-		var serviceMeta map[string]string
-		if meta != nil {
-			serviceMeta = meta[sDep.Name]
-		}
-
-		// Convert the hcat type to an HCL marshal-able object
-		s := newHealthService(sDep, serviceMeta)
-
-		f := hclwrite.NewEmptyFile()
-		gohcl.EncodeIntoBody(s, f.Body())
-		return strings.TrimSpace(string(f.Bytes()))
 	}
 }
